+++ luahelper-lsp/langserver/check/common/util.go
@@ -17,6 +17,8 @@ func GetTableAccessName(tabExp *ast.TableAccessExp) string {
 }
 
 // GetTableNameInfo 获取表表达式的表名
+// 例如table.a.b.c 取table
+// _G.table.a.b取table
 func GetTableNameInfo(tabExp *ast.TableAccessExp) (string, lexer.Location) {
 	for {
 		subExp, ok := tabExp.PrefixExp.(*ast.TableAccessExp)
@@ -25,7 +27,20 @@ func GetTableNameInfo(tabExp *ast.TableAccessExp) (string, lexer.Location) {
 		}
 		tabExp = subExp
 	}
-	return GetExpName(tabExp.PrefixExp), GetExpLoc(tabExp.PrefixExp)
+
+	keyName := GetExpName(tabExp.KeyExp)
+	if keyName == "tableB" {
+		keyName = "tableB"
+	}
+
+	//return GetExpName(tabExp.PrefixExp), GetExpLoc(tabExp.PrefixExp)
+	preName := GetExpName(tabExp.PrefixExp)
+	if preName == "!_G" {
+		return keyName, GetExpLoc(tabExp.KeyExp)
+	} else {
+		return GetSimpleValue(preName), GetExpLoc(tabExp.PrefixExp)
+	}
+
 }
 
 // GetExpType 获取exp的类型
@@ -49,6 +64,11 @@ func GetExpType(node ast.Exp) LuaType {
 		return LuaTypeTable
 
 	case *ast.BinopExp:
+		if exp.Op == lexer.TkOpEq || exp.Op == lexer.TkOpNe || exp.Op == lexer.TkOpLt || exp.Op == lexer.TkOpLe ||
+			exp.Op == lexer.TkOpGt || exp.Op == lexer.TkOpGe {
+			return LuaTypeBool
+		}
+
 		oneType := GetExpType(exp.Exp1)
 		if oneType != LuaTypeAll && oneType != LuaTypeRefer {
 			return oneType
@@ -74,6 +94,10 @@ func GetExpType(node ast.Exp) LuaType {
 		return LuaTypeRefer
 
 	case *ast.UnopExp:
+		if exp.Op == lexer.TkOpNot {
+			return LuaTypeBool
+		}
+
 		oneType := GetExpType(exp.Exp)
 		if oneType != LuaTypeAll {
 			return oneType
@@ -83,9 +107,6 @@ func GetExpType(node ast.Exp) LuaType {
 			return LuaTypeInter
 		}
 
-		if exp.Op == lexer.TkOpNot {
-			return LuaTypeBool
-		}
 	case *ast.NameExp:
 		return LuaTypeRefer
 	case *ast.ParensExp:
@@ -97,6 +118,35 @@ func GetExpType(node ast.Exp) LuaType {
 	}
 
 	return LuaTypeAll
+}
+
+// 把ast.Exp转换成annotateast.type的字符串
+func GetAnnTypeFromExp(referExp ast.Exp) string {
+	expType := GetExpType(referExp)
+
+	return GetAnnTypeFromLuaType(expType)
+}
+
+// 把LuaType转换成annotateast.type的字符串
+func GetAnnTypeFromLuaType(lua_type LuaType) string {
+	switch lua_type {
+	case LuaTypeNumber, LuaTypeInter, LuaTypeFloat:
+		return "number"
+	case LuaTypeString:
+		return "string"
+	case LuaTypeBool:
+		return "boolean"
+	case LuaTypeNil:
+		return "nil"
+	case LuaTypeFunc:
+		return "function"
+	case LuaTypeTable:
+		return "table"
+	case LuaTypeRefer:
+		return "LuaTypeRefer"
+	}
+
+	return "any"
 }
 
 // GetExpTypeString 获取ext的类型字符串
@@ -138,7 +188,7 @@ func GetLuaTypeString(luaType LuaType, referExp ast.Exp) string {
 	case LuaTypeNil:
 		return "any"
 	case LuaTypeBool:
-		return "bool"
+		return "boolean"
 	case LuaTypeInter:
 		if sufStr == "" {
 			return "number"