+++ luahelper-vscode/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 0.2.15 (Apr 9, 2022)
++ 对table的hover继续进行了优化
++ 修复了外网一处死循环崩溃
++ [Github](https://github.com/Tencent/LuaHelper)  https://github.com/Tencent/LuaHelper
+
+## 0.2.14 (Apr 6, 2022)
++ 修复了外网一处崩溃
++ 对table的hover展示进行了优化
++ [Github](https://github.com/Tencent/LuaHelper)  https://github.com/Tencent/LuaHelper
+
+## 0.2.12 (Mar 29, 2022)
++ 后台二进制采用go1.18版本编译
++ 优化了部分推导的类型
++ 增加了一个设置开关
++ 修复了外网的崩溃([issues:88](https://github.com/Tencent/LuaHelper/issues/88))
++ 优化了插件被激活的情况([issues:89](https://github.com/Tencent/LuaHelper/issues/89))
++ [Github](https://github.com/Tencent/LuaHelper)  https://github.com/Tencent/LuaHelper
+
 ## 0.2.11 (Mar 7, 2022)
 + 注解功能增加了alias功能
 + 增加显示了函数的参数类型与返回值